fix user permissions

+++ controllers/post.contoller.js
@@ -1,12 +1,24 @@
 const getStream = require("into-stream");
 const postModel = require("../models/post.model");
-const azureUpload = require("../utils/azure.upload");
+const ObjectId = require("mongoose").Types.ObjectId;
 const { checkId } = require("../models/user.model");
+const azureUpload = require("../utils/azure.upload");
 const { postValidation, checkImgFromat } = require("../utils/validation");
 
 exports.getPosts = async (req, res) => {
   try {
-    let posts = await postModel.getAllPosts();
+    let posts,
+      { userId } = req.query;
+
+    if (!userId) {
+      posts = await postModel.getAllPosts();
+    } else {
+      // Check user id
+      let checkResult = await checkId(userId);
+      if (checkResult) return res.status(400).send(checkResult);
+
+      posts = await postModel.getUserPosts(userId);
+    }
     res.status(200).json(posts);
   } catch (err) {
     res.status(500).send("Faild to get posts");
@@ -30,12 +42,14 @@ exports.getPost = async (req, res) => {
 
 exports.createNewPost = async (req, res) => {
   try {
-    let { title, body, createdBy } = req.body;
+    let createdBy = req.user._id;
+    let { title, body } = req.body;
 
-    if (!title || !body || !createdBy || !req.file) {
-      return res.status(400).send("title, body, createdBy, image are required");
+    if (!title || !body || !req.file) {
+      return res.status(400).send("title, body, image are required");
     }
 
+    // Check validation errors
     const validationResult = postValidation(req.body);
     if (validationResult) {
       return res.status(400).send(validationResult.details[0].message);
@@ -83,81 +97,66 @@ exports.createNewPost = async (req, res) => {
 exports.updatePost = async (req, res) => {
   try {
     let postImg,
-      id = req.params.id,
-      { title, body, createdBy } = req.body;
+      postId = req.params.id,
+      userId = req.user._id,
+      { title, body } = req.body;
 
     // Check post id
-    let result = await postModel.checkId(id);
+    let result = await postModel.checkId(postId);
     if (result) return res.status(400).send(result);
 
-    const validationResult = postValidation(req.body);
-    if (validationResult) {
-      return res.status(400).send(validationResult.details[0].message);
-    }
+    // Get old post details
+    let post = await postModel.getPostDetails(postId);
 
-    // Check user id
-    if (createdBy) {
-      let checkResult = await checkId(createdBy);
-      if (checkResult) return res.status(400).send(checkResult);
-    }
-
-    // Check if image is exist to update
-    if (req.file) {
-      // Check image format
-      let checkResult = checkImgFromat(req.file);
-      if (!checkResult) {
-        return res.status(400).send("Invalid image format");
+    if (ObjectId(post.createdBy._id).toString() === userId) {
+      // Check validation errors
+      const validationResult = postValidation(req.body);
+      if (validationResult) {
+        return res.status(400).send(validationResult.details[0].message);
       }
 
-      // Upload image to azure
-      postImg = azureUpload.getBlobName(req.file.originalname);
-      let stream = getStream(req.file.buffer);
-      let containerClient =
-        azureUpload.blobServiceClient.getContainerClient("postimgs");
-      let blockBlobClient = containerClient.getBlockBlobClient(postImg);
-
-      // Set buffer size to 4MB and max 20 buffers
-      let uploadOptions = { bufferSize: 4 * 1024 * 1024, maxBuffers: 20 };
-      await blockBlobClient.uploadStream(
-        stream,
-        uploadOptions.bufferSize,
-        uploadOptions.maxBuffers,
-        {
-          blobHTTPHeaders: { blobContentType: "image/jpeg" },
+      // Check if image is exist to update
+      if (req.file) {
+        // Check image format
+        let checkResult = checkImgFromat(req.file);
+        if (!checkResult) {
+          return res.status(400).send("Invalid image format");
         }
-      );
 
-      // Get old post details
-      let post = await postModel.getPostDetails(id);
+        // Upload image to azure
+        postImg = azureUpload.getBlobName(req.file.originalname);
+        let stream = getStream(req.file.buffer);
+        let containerClient =
+          azureUpload.blobServiceClient.getContainerClient("postimgs");
+        let blockBlobClient = containerClient.getBlockBlobClient(postImg);
 
-      // Delete old post image with its snapshots
-      blockBlobClient = containerClient.getBlockBlobClient(post.postImg);
-      blockBlobClient.delete({ deleteSnapshots: "include" });
+        // Set buffer size to 4MB and max 20 buffers
+        let uploadOptions = { bufferSize: 4 * 1024 * 1024, maxBuffers: 20 };
+        await blockBlobClient.uploadStream(
+          stream,
+          uploadOptions.bufferSize,
+          uploadOptions.maxBuffers,
+          {
+            blobHTTPHeaders: { blobContentType: "image/jpeg" },
+          }
+        );
 
-      // Update post
-      await postModel.updatePost(id, title, body, postImg, createdBy);
-      res.status(200).send("Successfully updated post");
-    } else {
-      //  If no exist image to update
-      await postModel.updatePost(id, title, body, postImg, createdBy);
+        // Delete old post image with its snapshots
+        blockBlobClient = containerClient.getBlockBlobClient(post.postImg);
+        blockBlobClient.delete({ deleteSnapshots: "include" });
 
-      res.status(200).send("Successfully updated post");
-    }
+        // Update post
+        await postModel.updatePost(postId, title, body, postImg);
+        res.status(200).send("Successfully updated post");
+      } else {
+        //  If no exist image to update
+        await postModel.updatePost(postId, title, body, postImg);
+
+        res.status(200).send("Successfully updated post");
+      }
+    } else return res.status(401).send("You can only update your posts");
   } catch (err) {
     res.status(500).send("Faild to update post");
-    console.log(err);
-  }
-};
-
-exports.deleteAllPosts = async (req, res) => {
-  try {
-    await postModel.deleteAllPosts();
-
-    //Delete storage container for post images from azure
-    await azureUpload.deleteContainerIfExist("postimgs");
-    res.status(200).send("Successfully deleted all posts");
-  } catch (err) {
-    res.status(500).send("Faild to delete posts");
     console.log(err);
   }
 };
@@ -165,22 +164,40 @@ exports.deleteAllPosts = async (req, res) => {
 exports.deletePostById = async (req, res) => {
   try {
     // Check id
-    let postId = req.params.id;
+    let postId = req.params.id,
+      userId = req.user._id;
     let checkResult = await postModel.checkId(postId);
     if (checkResult) return res.status(400).send(checkResult);
 
-    // Delete post from database
-    let post = await postModel.deletePostById(postId);
+    // Get post details
+    let postDetails = await postModel.getPostDetails(postId);
+    if (ObjectId(postDetails.createdBy._id).toString() === userId) {
+      // Delete post from database
+      let post = await postModel.deletePostById(postId);
 
-    //Delete post image from azure storage
-    let containerClient =
-      azureUpload.blobServiceClient.getContainerClient("postimgs");
-    let blockBlobClient = containerClient.getBlockBlobClient(post.postImg);
-    blockBlobClient.delete({ deleteSnapshots: "include" });
+      //Delete post image from azure storage
+      let containerClient =
+        azureUpload.blobServiceClient.getContainerClient("postimgs");
+      let blockBlobClient = containerClient.getBlockBlobClient(post.postImg);
+      blockBlobClient.delete({ deleteSnapshots: "include" });
 
-    res.status(200).send(`Successfully deleted post with id: ${postId}`);
+      res.status(200).send(`Successfully deleted post with id: ${postId}`);
+    } else return res.status(401).send("You can only delete your posts");
   } catch (err) {
     res.status(500).send("Faild to delete post");
     console.log(err);
   }
 };
+
+// exports.deleteAllPosts = async (req, res) => {
+//   try {
+//     await postModel.deleteAllPosts();
+
+//     //Delete storage container for post images from azure
+//     await azureUpload.deleteContainerIfExist("postimgs");
+//     res.status(200).send("Successfully deleted all posts");
+//   } catch (err) {
+//     res.status(500).send("Faild to delete posts");
+//     console.log(err);
+//   }
+// };
